Validator: expand info response Content-Type checks
http://iiif.io/api/image/2.1/index.html#information-request:

* When JSON-LD was not requested, the response MUST be
  `application/json`
* When JSON-LD is requested, the response must either be
  `application/ld+json` or `application/json` with a Link
  header specifying the standard context
