+++ implementations/validator/iiif_validator/tests/__init__.py
@@ -1,11 +1,12 @@
 
 __author__ = "IIIF Contributors"
 
-__all__ = ['baseurl_redirect', 'cors', 'format_conneg', 'format_error_random', 'format_gif', 
-  'format_jp2', 'format_jpg', 'format_pdf', 'format_png', 'format_tif', 'format_webp', 'id_basic', 
-  'id_error_escapedslash', 'id_error_random', 'id_error_unescaped', 'id_escaped', 'id_squares', 
-  'info_json', 'info_xml', 'jsonld', 'linkheader_canonical', 'linkheader_profile', 'quality_bitonal', 
-  'quality_color', 'quality_error_random', 'quality_grey', 'region_error_random', 'region_percent', 
-  'region_pixels', 'rot_error_random', 'rot_full_basic', 'rot_full_non90', 'rot_mirror', 
-  'rot_mirror_180', 'rot_region_basic', 'rot_region_non90', 'size_bwh', 'size_ch', 'size_error_random', 
+__all__ = ['baseurl_redirect', 'cors', 'format_conneg', 'format_error_random', 'format_gif',
+  'format_jp2', 'format_jpg', 'format_pdf', 'format_png', 'format_tif', 'format_webp', 'id_basic',
+  'id_error_escapedslash', 'id_error_random', 'id_error_unescaped', 'id_escaped', 'id_squares',
+  'info_default_response_content_type', 'info_explicit_json_ld_response_content_type',
+  'info_json', 'info_xml', 'linkheader_canonical', 'linkheader_profile', 'quality_bitonal',
+  'quality_color', 'quality_error_random', 'quality_grey', 'region_error_random', 'region_percent',
+  'region_pixels', 'rot_error_random', 'rot_full_basic', 'rot_full_non90', 'rot_mirror',
+  'rot_mirror_180', 'rot_region_basic', 'rot_region_non90', 'size_bwh', 'size_ch', 'size_error_random',
   'size_percent', 'size_region', 'size_up', 'size_wc', 'size_wh']
